feat: Herramientas de primer y segundo nivel para macOS (pantalla de alta resolución); efecto de desenfoque de barra

+++ src/tool/base/blurwidget.h
@@ -18,7 +18,7 @@ class BlurWidget : public QWidget
 public:
     explicit BlurWidget(QWidget *parent = nullptr);
     virtual ~BlurWidget() = default;
-    void setPixmap(const QPixmap& pix, double blurRadius);
+    void setPixmap(const QPixmap& pix, int blurRadius);
 
 protected:
     void paintEvent(QPaintEvent *event) override;
